% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{breaks}
\alias{breaks}
\title{Selects the breaks}
\usage{
breaks(df, col_name, n = 6, style = "quantile", pal = NULL,
  fixedBreaks = NULL, distribution = FALSE)
}
\arguments{
\item{df}{a data frame containing at least one numeric column}

\item{col_name}{name of the column containing the value on which the class
intervals should be calculate}

\item{n}{a numeric indicating the number of intervals to represent the data
(by default, \code{n = 6})}

\item{style}{a character value issued from the \code{classint} package, and
used to select a method for the different way of calculating the intervals
(by default \code{style = "quantile"})}

\item{pal}{the color for the distribution legend, by default \code{NULL}}

\item{fixedBreaks}{By default \code{NULL} but if a vector of value is
inputed, it will be used to specifen the breaks, need to used with the
parameter \code{style = "fixed"}}

\item{distribution}{if TRUE, print on the map, the distribution of the values
by intervals}
}
\value{
The data frame inputed with attributes \code{breaks} corresponding to
 the breaks value, returned invisibly.
}
\description{
This function selects the class intervals for the numeric column in a
data frame.
}
\examples{

# Example with data preparation
library(gdpm)
library(dplyr)
dengue <- getid(dengue, from = 1990, to = 1991)
dengue_prep <- filter(dengue, year == 1990, month == "September")

a <- breaks(dengue_prep, "incidence_dengue")
str(a)
# to extract the breaks value
attr(a, "breaks")

# to print the distribution
breaks(dengue_prep, "incidence_dengue", pal = heat.colors(6),
       distribution = TRUE)

# to change the style or number of breaks
breaks(dengue_prep, "incidence_dengue", n = 8, style = "fisher")
breaks(dengue_prep, "incidence_dengue", n = 8, style = "fisher",
       pal = heat.colors(8), distribution = TRUE)

# to use the "fixed" style
breaks(dengue_prep, "incidence_dengue", n = 6, style = "fixed",
       fixedBreaks = c(0, 50, 100, 500, 1000, 1500, 2000))
breaks(dengue_prep, "incidence_dengue", n = 6, style = "fixed",
       pal = heat.colors(6), distribution = TRUE,
       fixedBreaks = c(0, 50, 100, 500, 1000, 1500, 2000))

}
