% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infections_dates}
\alias{infections_dates}
\title{Vector of dates}
\format{A vector of 7018 dates.}
\usage{
infections_dates
}
\description{
A vector containing multiple dates in the format "%Y-%m-%d", used as example
for the \code{ll2incidence} function
}
\keyword{datasets}
