% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_prov.R
\name{merge_prov}
\alias{merge_prov}
\title{Merges provinces}
\usage{
merge_prov(df, sel = names(df), FUN = sum, from, to = "2017-12-31",
  diseases = NULL, df2 = NULL, args = NULL, FUN2 = sum)
}
\arguments{
\item{df}{A data frame containing at least the variables \code{province},
\code{year}, accept also the monthly data if the month are in the column
\code{month}.}

\item{sel}{A vector of character to select only the variable to merge.
By default, select all the variables.}

\item{FUN}{A function to apply on the data when merging the province
together. By default, \code{sum}.}

\item{from}{Initial date of the time range selected for the province
definition, of the class \code{Date} or \code{numeric}.}

\item{to}{Final date of the time range selected for the province
definition, of the class \code{Date} or \code{numeric}, by default
 \code{"2017-12-31"}}

\item{diseases}{A vector of character used to know which history of vietnam
has to be taken for the merging event as two diseases has different story.
Used if you want the same merging event as your infectious disease dataframe,
 for more details look at the /code{gdpm package}. By default, \code{NULL}}

\item{df2}{A data frame containing at least the variables \code{province},
\code{year}. Can be used to provide additional arguments through the
paramaters args by providing the name of the column(s) containing the data.}

\item{args}{A vector or list of additional arguments to pass to FUN}

\item{FUN2}{A function to apply on df2 when merging the province together,
to keep the same information at every step of the merging process. By
default \code{sum}}
}
\value{
A object of the same class as \code{df} in which all the provinces
that needed to be merged (according to the time range) are merged.
}
\description{
Tidy the data and merges data accordingly to a time range and by the
provinces concerned by a split/combined event and return a data frame for the
 time range imputed.
}
\examples{

# For all the following examples, we will use the data from the package gso.
library(gso) # for 'get_gso', 'pop_size'
mortality_rate <- get_gso("Infant mortality rate by province")

# if you want to have the data expressed by province, with the province's
# definition of 1992 in Vietnam:
merge_prov(mortality_rate, from = "1992-01-01")
# If you want the province's definition between 1992 and 2010 in Vietnam:
merge_prov(mortality_rate, from = 1992, to = 2010)

# You can change the function
merge_prov(mortality_rate, from = "1992", FUN = mean)

# You can also use weighted mean by providing the weighted in another
# data frame
pop_size <- gso::pop_size
merge_prov(mortality_rate, from = "1992-01-01", FUN = weighted.mean,
 df2 = pop_size, args = "total")

# You can define the merge_prov function only on certain columns
pop_info <- get_gso("Area, population and population density by province")
merge_prov(pop_info, sel = "average_population_thous_pers",
 from = 1992, FUN = weighted.mean,
 df2 = pop_size, args = "total")


}
