% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poseid.R
\docType{package}
\name{poseid}
\alias{poseid}
\alias{poseid-package}
\title{poseid: Population-based Observational Studies of the Epidemiology of Infectious Diseases}
\description{
The marc package provide functions of visualizing and analysing
population-based observational data in epidemiology.
}
